chore(refactor): resturctured project (#28)
* feat: implement core IR layer for hexagonal architecture

Add internal IR (Intermediate Representation) layer that sits between
AST and LLVM codegen, enabling backend independence.

- Create xvr_ir.h with IR type definitions (opcodes, types, values,
  instructions, basic blocks, functions, modules)
- Implement xvr_ir.c with creation and destruction functions
- Add proper memory management with malloc/calloc/free
- Include error handling (return NULL on allocation failure)
- Update CMakeLists.txt to include new source files

* refactor: move type system to core/types/

- Copy xvr_type.h and xvr_type.c to src/core/types/
- Update includes to use relative paths within core/
- Create redirect files in src/ that include the core versions
- Update CMakeLists.txt to compile from core/types/
- Fix header guard collision (XVR_TYPE_H -> XVR_CORE_TYPE_H)

* Move semantic analyzer to core/semantic/ directory

- Copy xvr_semantic.h and xvr_semantic.c to src/core/semantic/
- Update includes to use relative paths (../ast/, ../types/)
- Create redirect files in src/ that include from core/
- Update CMakeLists.txt to compile from core/semantic/

* chore(refactor): restructured project

this structured following hexagonal structured


* fix: accept max_len param / bounds check before copy

by change `xvr_ir_strdup()`


* fix: add bound check again for memcopy


* fix: using explicit alloc_size

make buffer size calculation clear to static analysis


* fix: create explicit buffer size validation


---------
